Add files via upload
Essentially my group's final project for CPR E 288 - Introduction to Embedded Systems.
This project features manual control via a platform named PuTTY, with the controls copying those of Video Game movements on a PC platform. 
The objective of the bot is to scan and void objects while it is controlled to find a pallet and then deliver the pallet to a "delivery bay" of choice, the final push is the final working copy of the project, with the other versions containing a sort of autonomous attempt to make this happen autonomously, but was ultimately scrapped due to timing constraints.